+++ sequentialPrime.h
@@ -1,6 +1,8 @@
 #ifndef _SEQUENTIAL_H
 #define _SEQUENTIAL_H
 
-int runSequential( unsigned int start, unsigned int end );
+#include "functions.h"
+
+int runSequential( ull start, ull end );
 
 #endif
